+++ buildSrc/src/main/kotlin/Project.kt
@@ -120,7 +120,7 @@ object Versions {
   const val mapboxGestures = "0.8.0"
   const val mapboxJavaServices = "5.4.1"
   const val mapboxBase = "0.8.0"
-  const val mapboxGlNative = "11.0.0-SNAPSHOT.0313T1613Z.f7fb332"
+  const val mapboxGlNative = "11.0.0-SNAPSHOT.0314T1520Z.e988934"
   const val mapboxCommon = "23.4.0-beta.1"
   const val androidxCore = "1.9.0" // last version compatible with kotlin 1.8.10
   const val androidxFragmentTesting = "1.5.0"