+++ sdk/src/testPublic/com/mapbox/maps/MapboxMapTest.kt
@@ -513,6 +513,13 @@ class MapboxMapTest {
   }
 
   @Test
+  fun cameraForCoordinateBoundsWithoutPadding() {
+    val bounds = mockk<CoordinateBounds>()
+    mapboxMap.cameraForCoordinateBounds(bounds)
+    verify { nativeMap.cameraForCoordinateBounds(bounds, null, null, null) }
+  }
+
+  @Test
   fun cameraForCoordinateBoundsOverloadBearing() {
     val bounds = mockk<CoordinateBounds>()
     mapboxMap.cameraForCoordinateBounds(bounds, bearing = 2.0)
@@ -593,6 +600,13 @@ class MapboxMapTest {
   }
 
   @Test
+  fun cameraForCoordinatesWithoutPadding() {
+    val points = mockk<List<Point>>()
+    mapboxMap.cameraForCoordinates(points)
+    verify { nativeMap.cameraForCoordinates(points, null, null, null) }
+  }
+
+  @Test
   fun cameraForCoordinatesOverloadBearing() {
     val points = mockk<List<Point>>()
     mapboxMap.cameraForCoordinates(points, bearing = 2.0)
@@ -619,6 +633,13 @@ class MapboxMapTest {
     val point = mockk<Point>()
     mapboxMap.cameraForGeometry(point)
     verify { nativeMap.cameraForGeometry(point, EdgeInsets(0.0, 0.0, 0.0, 0.0), null, null) }
+  }
+
+  @Test
+  fun cameraForGeometryWithoutPadding() {
+    val point = mockk<Point>()
+    mapboxMap.cameraForGeometry(point)
+    verify { nativeMap.cameraForGeometry(point, null, null, null) }
   }
 
   @Test